FILE-TYPES Cleanup
